+++ apps/backoffice-v2/src/domains/customer/fetchers.ts
@@ -43,6 +43,8 @@ const CustomerSchema = z.object({
       isExample: z.boolean().default(false),
       isDemoAccount: z.boolean().default(false),
       isDemoKybEnabled: z.boolean().default(false),
+      createKybAndOwnershipAssessment: z.boolean().default(true),
+      createIdentityVerification: z.boolean().default(true),
       demoAccessDetails: z
         .object({
           totalReports: z.number(),