+++ apps/kyb-app/CHANGELOG.md
@@ -17,6 +17,8 @@
   - @ballerine/blocks@0.2.46
   - @ballerine/common@0.9.105
   - @ballerine/ui@0.7.154
+- Updated dependencies
+  - @ballerine/common@0.9.105
   - @ballerine/workflow-browser-sdk@0.6.128
 
 ## 0.3.184